refactor(news): relocate ingest scripts under server package
Move scripts/news-ingest -> server/scripts/news-ingest so their bare
imports (drizzle-orm, @paperclipai/db) resolve from server/node_modules
on the pruned prod deploy (no root-level symlink hack). Run via
node cli/node_modules/tsx/dist/cli.mjs server/scripts/news-ingest/<x>.ts.
